rename container to features
