+++ client/pages/customer.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import withApollo from '../src/lib/withApollo'
-import Customer from '../src/container/customer'
+import Customer from '../src/features/customer'
 
 const CustomerPage = () => <Customer />
