+++ client/pages/index.js
@@ -1,6 +1,7 @@
 import React from 'react'
-import Board from '../src/container/board'
+
 import withApollo from '../src/lib/withApollo'
+import Board from '../src/features/board'
 
 const Home = () => <Board />
