+++ client/pages/product.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import withApollo from '../src/lib/withApollo'
-import Product from '../src/container/product'
+import Product from '../src/features/product'
 
 const ProductPage = () => <Product />
